What about this to make it an installable module?
