+++ ovro_data_recorder/data/logo.png
