+++ scripts/dr_visibilities.py
@@ -15,7 +15,7 @@ import threading
 from functools import reduce
 from datetime import datetime, timedelta
 
-from gridder import WProjection
+from ovro_data_recorder.gridder import WProjection
 from scipy.stats import scoreatpercentile as percentile
 
 from lwa_antpos.station import ovro
@@ -23,14 +23,14 @@ from lwa_antpos.station import ovro
 from mnc.common import *
 from mnc.mcs import ImageMonitorPoint, MultiMonitorPoint, Client
 
-from reductions import *
-from operations import FileOperationsQueue
-from monitoring import GlobalLogger
-from control import VisibilityCommandProcessor
-from lwams import get_zenith_uvw
-from version import version as repo_version
+from ovro_data_recorder.reductions import *
+from ovro_data_recorder.operations import FileOperationsQueue
+from ovro_data_recorder.monitoring import GlobalLogger
+from ovro_data_recorder.control import VisibilityCommandProcessor
+from ovro_data_recorder.lwams import get_zenith_uvw
+from ovro_data_recorder.version import version as repo_version
 
-from xengine_fast_control import FastStation
+from ovro_data_recorder.xengine_fast_control import FastStation
 
 from bifrost.address import Address
 from bifrost.udp_socket import UDPSocket
@@ -742,7 +742,7 @@ class ImageOp(object):
         draw.text((835, 30), '|V|', font = font, fill = '#FFFFFF')
         
         ## Logo-ize
-        logo = PIL.Image.open(os.path.join(BASE_PATH, 'logo.png'))
+        logo = PIL.Image.open(os.path.join(BASE_PATH, 'data', 'logo.png'))
         logo_img = logo.getchannel('A')
         im.paste(logo_img, (5, 385))
         logo.close()
