+++ setup.py
@@ -1,28 +1,84 @@
-import os
-import numpy
-import tempfile
-import subprocess
-from setuptools import setup, Extension, find_packages
-from distutils import log
-from distutils.command.install import install
+import glob
+from subprocess import check_call, check_output, CalledProcessError
+from setuptools import setup, Extension, find_namespace_packages
+
+try:
+    import numpy as np
+except Exception as e:
+    raise RuntimeError(f"numpy is required to run setup.py: {str(e)}")
 
 
-class dummy_install(install):
-    """Dummay install method that doesn't let you install."""
-    def finalize_options(self, *args, **kwargs):
-        raise RuntimeError("This is a dummy package that cannot be installed")
+def get_version():
+    """Determine a version based on the git repo info."""
+
+    # Query the git repo info.  If this fails make a note of it and yield an
+    # unknown version
+    version = 'unknown'
+    try:
+        git_branch = check_output(['git', 'branch', '--show-current'],
+                                  cwd=REPO_PATH)
+        git_branch = git_branch.decode().strip().rstrip()
+        git_hash = check_output(['git', 'log', '-n', '1', '--pretty=format:%H'],
+                                cwd=REPO_PATH)
+        git_hash = git_hash.decode().strip().rstrip()
+        
+        git_dirty = 0
+        try:
+            check_call(['git', 'diff-index', '--quiet', '--cached', 'HEAD', '--'],
+                       cwd=REPO_PATH)
+        except CalledProcessError:
+            git_dirty += 1
+        try:
+            check_call(['git', 'diff-files', '--quiet'],
+                       cwd=REPO_PATH)
+        except CalledProcessError:
+            git_dirty += 1
+            
+        version = git_branch+'@'+git_hash[:7]
+        if git_dirty > 0:
+            version += ' (dirty)'
+            
+    except CalledProcessError as e:
+        print(f"Failed to determine git repo versioning - {str(e)}")
+        
+    return version
+
+
+def write_version_info():
+    """Write the version info to a module in ovro_data_recorder."""
+    
+    odrVersion = get_version()
+    
+    with open('ovro_data_recorder/version.py', 'w') as fh:
+        fh.write(f"""# This file is automatically generated by setup.py
+
+version = '{odrVersion}'
+full_version = '{odrVersion}'
+""")
 
 
 ExtensionModules = [Extension('gridder', ['gridder.cpp',],
-                              include_dirs=[numpy.get_include()],
+                              include_dirs=[np.get_include()],
                               libraries=['m', 'fftw3f'],
                               extra_compile_args=['-DNPY_NO_DEPRECATED_API=NPY_1_7_API_VERSION',]),]
 
 
+# Update the version information
+write_version_info()
+
+
 setup(
-    cmdclass = {'install': dummy_install}, 
-    name = 'dummy_package',
-    version = '0.0',
-    description = 'This is a dummy package to help build the gridder extensions',
-    ext_modules = ExtensionModules
+    name = 'ovro_data_recorder',
+    version = get_version(),
+    description = 'Data recorder software for the Owens Valley Long Wavelength Array station',
+    license='BSD3',
+    packages=find_namespace_packages(),
+    scripts=glob.glob('scripts/*.py')
+    setup_requires = ['numpy>=1.7'],
+    install_requires = ['astropy', 'casacore', 'numpy', 'scipy', 'h5py', 'PIL', 'etcd3',
+                        'bifrost', 'mnc', 'lwa_antpos', 'lwa352_pipeline_control'],
+    include_package_data = True,  
+    ext_package = 'ovro_data_recorder', 
+    ext_modules = ExtensionModules,
+    zip_safe = False
 )
